[CI] publish: allow pushing a release to an existing tag
It is necessary to set the tag before uploading to a repository that
does not already have it. But when it does, there is no need. This is
the case for instance in the forgejo integration repository.

It is however best to verify the SHA of the tag is as expected to not
attempt to override a release associated with a different SHA,
presumably from a previous run.