[UI] Change how identities are shown in commit page
The username is now always shown next to full names. The commit
information should use the full name associated with the commit and the
signature information should use the full name associated with
the profile, provided that it exists.

This should remove any confusion if the author decides to use different
names under the same account.