[PRIVACY] Disable Gravatar by default
- Resolves #259
- Replaces #716 (thanks @silentcode !)
- The main toggle _for now_ (due to the still weird settings migration
to the database) is `GetDefaultDisableGravatar`, because it's used
`loadPictureFrom` which in turn sets the default value for the install
page which in turn is the explicit value that Forgejo will use to decide
to enable or disable Gravatar after the install page.
- Tests still rely on the fact that gravatar is enabled by default for
some tests, so there's a small hack in `MainTest` for that. This is not
enabled per-tests as this would otherwise just cause random test
failures for when rebasing happens.
- Unit tests ensures this value by default stays false.