[GITEA] avoid hard-coding height in language dropdown menu
This commit removes the hard-coded height of 500px, using that as a
max-height instead. The height of items in the dropdown menu, assuming
a default font size of 16px, is 36px, so the old CSS would cause
overly large dropdown menus in instances of Forgejo where less than 14
languages are offered.