Add support for different Git object formats
- Git [has committed](https://git-scm.com/docs/hash-function-transition)
to migrate away from SHA1 to a new hash algorithm in light of the
[shattered](https://shattered.io/) attack.
- Git made the good decision to create a new concept called Object
Format along with this change, which represents how objects are stored
and is currently tied to a hash algorithm. If in the future another
attack becomes feasible against a current hash algorithm, it will be
easier to go away because of the introduction of that abstraction of
modularization.
- This patch follows the same abstraction by moving all "old" Object
Format code to this same abstraction and thus closely follows what Git
does and therefore makes it much easier to add new object formats if Git
decides to add them in the hopefully far future.
- Because the sha256 object format is still experimental and not ready
according to the Git developers, this patch does not recommend using it
and is guarded behind a config flag.
- Personally, I hope that by merging this patch into Forgejo and Gitea's
codebase that the hash function transition will get some love again from
the Git developers, because right now it seems to be getting stale
because of, among other things, because forges don't support this new
object format.[^1]

[^1]: https://lwn.net/Articles/898522/