[GITEA] Don't post-proccess issue index in documents
- Currently the post processing will transform all issue indexes (such
as  `#6`) into an clickable link.
- This makes sense in an situation like issues or PRs, where referencing
to other issues is quite common and only referencing to their issue
indexing is an handy and efficient way to do it.
- Currently this is also run for documents (which is the user profile
and viewing rendered files), but in those situations it's less common to
reference issues by their index and instead could mean something else.
- This patch disables this post processing for issue index for
documents. Which matches Github's behavior.
- Added unit tests.
- Resolves https://codeberg.org/Codeberg/Community/issues/1120