[GITEA] Fix checkbox state not being saved on issue forms
- When creating a issue with the issue form feature, issue forms has
different inputs: dropdown, checkbox, textarea and a simple input. After
you've checked some checkboxes and filled in some text it's the expected
behavior that these states are actually shown in the created issue.
- This was not the case for checked checkboxes, as due to a regression
in the template the id of the checkboxes section wasn't included in the
checkbox element, therefor Forgejo didn't knew the checkbox was checked.
- Added an issue form to `repo59`.
- Added integration test for the rendering of the template.
- Regression of a6175b01d92a75dcbadbbfc6782a486636fd62a2
- Resolves https://codeberg.org/forgejo/forgejo/issues/1314