[GITEA] Remove tabindex on create issue button
- The tab order is important for those using a11y tools or those not
wanting to leave the keyboard to navigate around the site. Browsers can
usually figure out the best tab order if semantic HTML is being used, if
not, you can manually specify the tabindex on elements.
- This was the case in Gogs, however throughout the years this situation
has been improved and a lot of components are using semantic HTML and
therefor specifying tabindex isn't needed anymore and in this case
actually harms the tab order as the create issue button was the second
element in the tab order and the issue title the first element, which
isn't consistent with other pages where the header is the first element
tab order.