doctor: Let `doctor check` work as in v1.21+
In Forgejo v1.21+, the way to invoke a doctor check is `forgejo doctor
check`. However, v1.20 does not recognise the `check` command, so trying
to run the same doctor command one would run on 1.21+ will run the
default checks instead.

This patch adjusts the doctor command so that it recognises both the
v1.20 invocation, and the v1.21 style too.

Addresses one part of #1905.

Signed-off-by: Gergely Nagy <forgejo@gergo.csillger.hu>