[GITEA] Correct ErrNotExist condition in packages
- Currently when packages need to rebuild things, they would first
remove the original file, however it's possible that this nog longer
exist and thus the deletion should be skipped. The current condition
didn't do this and instead only prevented the function from returning
that the package file didn't exist, which caused NPE.
- Modify the behavior such that when the package doesn't exist it will
return `nil` in the case of Alpine package as there's nothing to
rebuild then. Continue to the next file to delete in the case of Debian
package.
- Resolves #1909