db: Automatically adjust database props on init
When initializing an empty MySQL/MariaDB database, ensure that the
character set and collation is set to the desired values.

With these set, creating a table will inherit these settings, unless
table creation specifies a different charset or collate function. This
is the reason why `setting.Database.MysqlCharset` was removed: that
forced table creation to explicitly set a charset, rather than inherit
the database's default, and in doing so, also changed the collate
function to the charset's default (which may - and usually is -
different from the one we want).

Signed-off-by: Gergely Nagy <forgejo@gergo.csillger.hu>