[GITEA] Require password confirmation to disable 2FA
- Add a password confirmation to disable 2FA on a user account. This is
to avoid anyone being able to disable 2FA silently on a logged-in
account.
- Ideally a security key is asked, if configured, but in the current
authentication schema a password is still the first-party authentication
even if a security key is setup. Even more ideally a uniform mechanism
is built that asks for 2FA/Security key/Password and than allow any
dangerous action to happen (such as deleting repository or your
account). For now a poor mans solution, equally secure as 'Delete your account'.
- Adds integration test.